- Remove FeederFacadeImpl.kt
