+++ feeder-web/src/main/kotlin/ru/timmson/feeder/service/FeederFacade.kt
@@ -1,8 +1,44 @@
 package ru.timmson.feeder.service
 
-interface FeederFacade {
+import org.springframework.stereotype.Service
+import ru.timmson.feeder.bot.BotService
+import ru.timmson.feeder.common.FeederConfig
+import ru.timmson.feeder.common.logger
+import ru.timmson.feeder.stock.service.StockService
 
-    fun sendStocksToOwner()
+@Service
+class FeederFacade(
+    private val feederConfig: FeederConfig,
+    private val stockService: StockService,
+    private val botService: BotService
+) {
 
-    fun sendStocksToChannel()
+    private val log = logger<FeederFacade>()
+
+    private val stocks = mapOf(
+        "usd" to "💰",
+        "imoex" to "🇷🇺",
+        "mredc" to "🏡",
+        "spx" to "🇺🇸",
+        "shcomp" to "🇨🇳"
+    )
+
+    fun sendStocksToOwner() =
+        sendStocks { message: String -> botService.sendMessageToOwner(message) }
+
+    fun sendStocksToChannel() =
+        sendStocks { message: String -> botService.sendMessage(feederConfig.stockChannelId, message) }
+
+
+    private fun sendStocks(send: (String) -> Unit) {
+        log.info("Entering sendStocks() ...")
+
+        val message = stockService.findAll().joinToString(", ") {
+            stocks.getOrDefault(it.ticker, "") + it.price
+        }
+
+        send(message)
+
+        log.info("Leaving sendStocks(...)")
+    }
 }
