+++ feeder-web/src/test/kotlin/ru/timmson/feeder/service/FeederFacadeImplShould.kt
@@ -9,29 +9,29 @@ import org.mockito.Mockito.`when`
 import org.mockito.junit.jupiter.MockitoExtension
 import org.mockito.kotlin.eq
 import org.mockito.kotlin.verify
-import ru.timmson.feeder.bot.BotService
+import ru.timmson.feeder.bot.BotServiceImpl
 import ru.timmson.feeder.common.FeederConfig
 import ru.timmson.feeder.stock.model.Stock
-import ru.timmson.feeder.stock.service.StockService
+import ru.timmson.feeder.stock.service.StockServiceImpl
 import java.math.BigDecimal
 
 @ExtendWith(MockitoExtension::class)
 class FeederFacadeImplShould {
 
-    private lateinit var feederFacade: FeederFacadeImpl
+    private lateinit var feederFacade: FeederFacade
 
     private lateinit var feederConfig: FeederConfig
 
     @Mock
-    private lateinit var stockService: StockService
+    private lateinit var stockService: StockServiceImpl
 
     @Mock
-    private lateinit var botService: BotService
+    private lateinit var botService: BotServiceImpl
 
     @BeforeEach
     fun setUp() {
         feederConfig = FeederConfig()
-        feederFacade = FeederFacadeImpl(feederConfig, stockService, botService)
+        feederFacade = FeederFacade(feederConfig, stockService, botService)
     }
 
     @Test
